refactor dirs
